Скрипт разбит на модули
